Compare file path instead of references when resolving objects.
This fixes a bug in the `link.resolve` function.  For me (perhaps due to changes in the most recent metalsmith version or perhaps due to other plugins I'm using, I'm not sure) the compare by reference method failed to resolve any object-based links as the file objects passed to the plugin seem to differ significantly from objects passed to `resolve`.  Matching on just the path attribute seems to work though (although I wonder if there was a reason that wasn't done originally?).